MybibleAI/backend/requirements.txt
Django>=4.2.2
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
